+++ src/main/java/blusunrize/immersiveengineering/common/gui/ToolboxMenu.java
@@ -16,6 +16,8 @@ import blusunrize.immersiveengineering.common.gui.IESlot.ICallbackContainer;
 import blusunrize.immersiveengineering.common.items.ToolboxItem;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.player.Inventory;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.inventory.ClickType;
 import net.minecraft.world.inventory.MenuType;
 import net.minecraft.world.inventory.Slot;
 import net.minecraft.world.item.ItemStack;
@@ -25,6 +27,8 @@ import net.minecraftforge.items.ItemStackHandler;
 
 public class ToolboxMenu extends IEContainerMenu implements ICallbackContainer
 {
+	protected int blockedSlot = -1;
+
 	public static ToolboxMenu makeFromBE(
 			MenuType<?> type, int id, Inventory invPlayer, ToolboxBlockEntity be
 	)
@@ -40,12 +44,24 @@ public class ToolboxMenu extends IEContainerMenu implements ICallbackContainer
 				itemCtx(type, id, invPlayer, slot, stack),
 				invPlayer,
 				CapabilityUtils.getPresentCapability(stack, ForgeCapabilities.ITEM_HANDLER)
-		);
+		).setBlockedSlot(invPlayer);
 	}
 
 	public static ToolboxMenu makeClient(MenuType<?> type, int id, Inventory invPlayer)
 	{
 		return new ToolboxMenu(clientCtx(type, id), invPlayer, new ItemStackHandler(ToolboxItem.SLOT_COUNT));
+	}
+
+	public static ToolboxMenu makeClientItem(MenuType<?> type, int id, Inventory invPlayer)
+	{
+		return new ToolboxMenu(clientCtx(type, id), invPlayer, new ItemStackHandler(ToolboxItem.SLOT_COUNT))
+				.setBlockedSlot(invPlayer);
+	}
+
+	public ToolboxMenu setBlockedSlot(Inventory invPlayer)
+	{
+		this.blockedSlot = (invPlayer.selected+27+ToolboxItem.SLOT_COUNT);
+		return this;
 	}
 
 	public ToolboxMenu(MenuContext ctx, Inventory inventoryPlayer, IItemHandler inv)
@@ -100,8 +116,22 @@ public class ToolboxMenu extends IEContainerMenu implements ICallbackContainer
 	}
 
 	@Override
+	public void clicked(int par1, int par2, ClickType par3, Player par4EntityPlayer)
+	{
+		if(blockedSlot >= 0&&(par1==this.blockedSlot||(par3==ClickType.SWAP&&par2==par4EntityPlayer.getInventory().selected)))
+			return;
+		super.clicked(par1, par2, par3, par4EntityPlayer);
+	}
+
+	@Override
 	public boolean canTake(ItemStack stack, int slotNumer, Slot slotObject)
 	{
 		return true;
+	}
+
+	@Override
+	public boolean isValidSlotIndex(int slotIndex)
+	{
+		return blockedSlot < 0||slotIndex!=this.blockedSlot;
 	}
 }