+++ changelog.md
@@ -2,6 +2,7 @@
 - Fix block breaking overlay on multiblocks (BluSunrize)
 - Fix tools playing a re-equip animation when they shouldn't (BluSunrize)
 - Fix windmills having their insides turn white when far away (BluSunrize)
+- Fix Toolbox being movable while opened (BluSunrize)
 
 ##### Version 1.20.1-10.1.0-171
 - Includes all features from 9.2.4